First tested and working version.
Bot that parses messages from a given website, and distributes them to
the users.
Handling of user commands, un/subscribing and starting/stoping.
Handling of changes in the website, and distribution of new matches.
Easily customisable, I suppose.
(Wanted to make a Telegram bot for a long time, this one is actually usefuseful for me)